hw1-classic-cv/requirements.txt
numpy>=1.20.0
pandas>=1.3.0
opencv-python>=4.5.0
scikit-learn>=0.24.0
torch>=1.10.0
torchvision>=0.11.0
tqdm>=4.62.0
PyYAML>=6.0
Pillow>=8.3.0
imagehash>=4.2.0